chatbot-app/requirements.txt
Flask==3.0.0
flask-login==0.6.3
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
email-validator==2.1.0
Werkzeug==3.0.1
pytz==2023.3
google-generativeai==0.3.2
gunicorn==21.2.0
requests==2.31.0

